Move software skinning to pl3DPipeline
For DX and GL at least, Metal does its own thing using simd directly.
