+++ Sources/Plasma/FeatureLib/pfGLPipeline/CMakeLists.txt
@@ -18,7 +18,7 @@ plasma_library(pfGLPipeline
     FOLDER FeatureLib
     SOURCES ${pfGLPipeline_SOURCES} ${pfGLPipeline_HEADERS}
 )
-plasma_target_simd_sources(pfGLPipeline SSE3 SSE41 plGLPipeline.cpp)
+plasma_target_simd_sources(pfGLPipeline SSE3 plGLPipeline.cpp)
 target_link_libraries(pfGLPipeline
     PUBLIC
         CoreLib
