+++ Sources/Plasma/FeatureLib/pfGLPipeline/plGLPipeline.h
@@ -65,10 +65,6 @@ class plGLPipeline : public pl3DPipeline<plGLDevice>
     friend class plGLPlateManager;
     friend class plGLDevice;
 
-protected:
-    typedef void(*blend_vert_buffer_ptr)(const plSpan*, hsMatrix44*, int, const uint8_t*, uint8_t , uint32_t, uint8_t*, uint32_t, uint32_t, uint16_t);
-    static hsCpuFunctionDispatcher<blend_vert_buffer_ptr> blend_vert_buffer;
-
 public:
     plGLPipeline(hsDisplayHndl display, hsWindowHndl window, const hsG3DDeviceModeRecord *devMode);
     virtual ~plGLPipeline();
@@ -83,7 +79,6 @@ public:
 
     /*** VIRTUAL METHODS ***/
     bool PreRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr=nullptr) override;
-    bool PrepForRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr=nullptr) override;
     plTextFont* MakeTextFont(ST::string face, uint16_t size) override;
     void CheckVertexBufferRef(plGBufferGroup* owner, uint32_t idx) override;
     void CheckIndexBufferRef(plGBufferGroup* owner, uint32_t idx) override;
@@ -134,28 +129,6 @@ protected:
      * buffers.
      */
     bool ICheckDynBuffers(plDrawableSpans* drawable, plGBufferGroup* group, const plSpan* span);
-
-    /**
-     * Emulate matrix palette operations in software.
-     *
-     * The big difference between the hardware and software versions is we only
-     * want to lock the vertex buffer once and blend all the verts we're going
-     * to in software, so the vertex blend happens once for an entire drawable.
-     * In hardware, we want the opposite, to break it into managable chunks,
-     * manageable meaning few enough matrices to fit into hardware registers.
-     * So for hardware version, we set up our palette, draw a span or few,
-     * setup our matrix palette with new matrices, draw, repeat.
-     */
-    bool ISoftwareVertexBlend(plDrawableSpans* drawable, const std::vector<int16_t>& visList);
-
-    /**
-     * Given a pointer into a buffer of verts that have blending data in the
-     * plVertCoder format, blends them into the destination buffer given
-     * without the blending info.
-     */
-    void IBlendVertsIntoBuffer(const plSpan* span, hsMatrix44* matrixPalette, int numMatrices, const uint8_t* src, uint8_t format, uint32_t srcStride, uint8_t* dest, uint32_t destStride, uint32_t count, uint16_t localUVWChans) {
-        blend_vert_buffer.call(span, matrixPalette, numMatrices, src, format, srcStride, dest, destStride, count, localUVWChans);
-    };
 
 private:
     static plGLEnumerate enumerator;
