+++ Sources/Plasma/PubUtilLib/plPipeline/pl3DPipeline.h
@@ -82,6 +82,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 class plDebugTextManager;
 class plPlateManager;
 
+plProfile_Extern(PrepDrawable);
 plProfile_Extern(RenderScene);
 plProfile_Extern(VisEval);
 plProfile_Extern(VisSelect);
@@ -100,6 +101,10 @@ plProfile_Extern(LightChar);
 plProfile_Extern(LightActive);
 plProfile_Extern(FindLightsFound);
 plProfile_Extern(FindLightsPerm);
+plProfile_Extern(AvatarSort);
+plProfile_Extern(AvatarFaces);
+plProfile_Extern(Skin);
+plProfile_Extern(NumSkin);
 
 static const float kPerspLayerScale  = 0.00001f;
 static const float kPerspLayerScaleW = 0.001f;
@@ -263,6 +268,8 @@ protected:
     uint16_t                                fAvRTWidth;
     uint32_t                                fAvNextFreeRT;
 
+    typedef void (*blend_vert_buffer_ptr)(const plSpan*, hsMatrix44*, int, const uint8_t*, uint8_t , uint32_t, uint8_t*, uint32_t, uint32_t, uint16_t);
+    static hsCpuFunctionDispatcher<blend_vert_buffer_ptr> fBlendVertBuffer;
 
 public:
     pl3DPipeline(const hsG3DDeviceModeRecord* devModeRec);
@@ -274,7 +281,7 @@ public:
 
     /*** VIRTUAL METHODS ***/
     //virtual bool PreRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr=nullptr) = 0;
-    //virtual bool PrepForRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr=nullptr) = 0;
+    bool PrepForRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr=nullptr) override;
 
 
     /**
@@ -912,6 +919,54 @@ protected:
 
     /** pass the current local to world tranform on to the device. */
     void ILocalToWorldToDevice();
+
+    /**
+     * Sorts the avatar geometry for display.
+     *
+     * We handle avatar sort differently from the rest of the face sort. The
+     * reason is that within the single avatar index buffer, we want to only
+     * sort the faces of spans requesting a sort, and sort them in place.
+     *
+     * Contrast that with the normal scene translucency sort. There, we sort
+     * all the spans in a drawble, then we sort all the faces in that drawable,
+     * then for each span in the sorted span list, we extract the faces for
+     * that span appending onto the index buffer. This gives great efficiency
+     * because only the visible faces are sorted and they wind up packed into
+     * the front of the index buffer, which permits more batching. See
+     * plDrawableSpans::SortVisibleSpans.
+     *
+     * For the avatar, it's generally the case that all the avatar is visible
+     * or not, and there is only one material, so neither of those efficiencies
+     * is helpful. Moreover, for the avatar the faces we want sorted are a tiny
+     * subset of the avatar's faces. Moreover, and most importantly, for the
+     * avatar, we want to preserve the order that spans are drawn, so, for
+     * example, the opaque base head will always be drawn before the
+     * translucent hair fringe, which will always be drawn before the pink
+     * clear plastic baseball cap.
+     */
+    bool IAvatarSort(plDrawableSpans* d, const std::vector<int16_t>& visList);
+
+    /**
+     * Emulate matrix palette operations in software.
+     *
+     * The big difference between the hardware and software versions is we only
+     * want to lock the vertex buffer once and blend all the verts we're going
+     * to in software, so the vertex blend happens once for an entire drawable.
+     * In hardware, we want the opposite, to break it into managable chunks,
+     * manageable meaning few enough matrices to fit into hardware registers.
+     * So for hardware version, we set up our palette, draw a span or few,
+     * setup our matrix palette with new matrices, draw, repeat.
+     */
+    bool ISoftwareVertexBlend(plDrawableSpans* drawable, const std::vector<int16_t>& visList);
+
+    /**
+     * Given a pointer into a buffer of verts that have blending data in the
+     * plVertCoder format, blends them into the destination buffer given
+     * without the blending info.
+     */
+    void IBlendVertsIntoBuffer(const plSpan* span, hsMatrix44* matrixPalette, int numMatrices, const uint8_t* src, uint8_t format, uint32_t srcStride, uint8_t* dest, uint32_t destStride, uint32_t count, uint16_t localUVWChans) {
+        fBlendVertBuffer.call(span, matrixPalette, numMatrices, src, format, srcStride, dest, destStride, count, localUVWChans);
+    };
 };
 
 
@@ -1000,6 +1055,39 @@ pl3DPipeline<DeviceType>::~pl3DPipeline()
 
     IClearClothingOutfits(&fClothingOutfits);
     IClearClothingOutfits(&fPrevClothingOutfits);
+}
+
+
+template<class DeviceType>
+bool pl3DPipeline<DeviceType>::PrepForRender(plDrawable* drawable, std::vector<int16_t>& visList, plVisMgr* visMgr)
+{
+    plProfile_TimingGuard(PrepDrawable);
+
+    plDrawableSpans* ice = plDrawableSpans::ConvertNoRef(drawable);
+    if (!ice)
+        return false;
+
+    // Find our lights
+    ICheckLighting(ice, visList, visMgr);
+
+    // Sort our faces
+    if (ice->GetNativeProperty(plDrawable::kPropSortFaces))
+        ice->SortVisibleSpans(visList, this);
+
+    // Prep for render. This is gives the drawable a chance to
+    // do any last minute updates for its buffers, including
+    // generating particle tri lists.
+    ice->PrepForRender(this);
+
+    // Any skinning necessary
+    if (!ISoftwareVertexBlend(ice, visList))
+        return false;
+
+    // Avatar face sorting happens after the software skin.
+    if (ice->GetNativeProperty(plDrawable::kPropPartialSort))
+        IAvatarSort(ice, visList);
+
+    return true;
 }
 
 
@@ -1971,6 +2059,226 @@ void pl3DPipeline<DeviceType>::ILocalToWorldToDevice()
 {
     fDevice.SetLocalToWorldMatrix(fView.GetLocalToWorld());
     fView.fXformResetFlags &= ~fView.kResetL2W;
+}
+
+
+struct plSortFace
+{
+    uint16_t      fIdx[3];
+    float    fDist;
+};
+
+struct plCompSortFace
+{
+    bool operator()( const plSortFace& lhs, const plSortFace& rhs) const
+    {
+        return lhs.fDist > rhs.fDist;
+    }
+};
+
+template <class DeviceType>
+bool pl3DPipeline<DeviceType>::IAvatarSort(plDrawableSpans* d, const std::vector<int16_t>& visList)
+{
+    plProfile_TimingGuard(AvatarSort);
+    for (int16_t visIdx : visList)
+    {
+        hsAssert(d->GetSpan(visIdx)->fTypeMask & plSpan::kIcicleSpan, "Unknown type for sorting faces");
+
+        plIcicle* span = (plIcicle*)d->GetSpan(visIdx);
+
+        if (span->fProps & plSpan::kPartialSort) {
+            hsAssert(d->GetBufferGroup(span->fGroupIdx)->AreIdxVolatile(), "Badly setup buffer group - set PartialSort too late?");
+
+            const hsPoint3 viewPos = GetViewPositionWorld();
+
+            plGBufferGroup* group = d->GetBufferGroup(span->fGroupIdx);
+
+            typename DeviceType::VertexBufferRef* vRef = static_cast<typename DeviceType::VertexBufferRef*>(group->GetVertexBufferRef(span->fVBufferIdx));
+
+            const uint8_t* vdata = vRef->fData;
+            const uint32_t stride = vRef->fVertexSize;
+
+            const int numTris = span->fILength/3;
+
+            static std::vector<plSortFace> sortScratch;
+            sortScratch.resize(numTris);
+
+            plProfile_IncCount(AvatarFaces, numTris);
+
+            // Have three very similar sorts here, differing only on where the "position" of
+            // each triangle is defined, either as the center of the triangle, the nearest
+            // point on the triangle, or the farthest point on the triangle.
+            // Having tried all three on the avatar (the only thing this sort is used on),
+            // the best results surprisingly came from using the center of the triangle.
+            uint16_t* indices = group->GetIndexBufferData(span->fIBufferIdx) + span->fIStartIdx;
+            int j;
+            for( j = 0; j < numTris; j++ )
+            {
+#if 1 // TRICENTER
+                uint16_t idx = *indices++;
+                sortScratch[j].fIdx[0] = idx;
+                hsPoint3 pos = *(hsPoint3*)(vdata + idx * stride);
+
+                idx = *indices++;
+                sortScratch[j].fIdx[1] = idx;
+                pos += *(hsPoint3*)(vdata + idx * stride);
+
+                idx = *indices++;
+                sortScratch[j].fIdx[2] = idx;
+                pos += *(hsPoint3*)(vdata + idx * stride);
+
+                pos *= 0.3333f;
+
+                sortScratch[j].fDist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+#elif 0 // NEAREST
+                uint16_t idx = *indices++;
+                sortScratch[j].fIdx[0] = idx;
+                hsPoint3 pos = *(hsPoint3*)(vdata + idx * stride);
+                float dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                float minDist = dist;
+
+                idx = *indices++;
+                sortScratch[j].fIdx[1] = idx;
+                pos = *(hsPoint3*)(vdata + idx * stride);
+                dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                if( dist < minDist )
+                    minDist = dist;
+
+                idx = *indices++;
+                sortScratch[j].fIdx[2] = idx;
+                pos = *(hsPoint3*)(vdata + idx * stride);
+                dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                if( dist < minDist )
+                    minDist = dist;
+
+                sortScratch[j].fDist = minDist;
+#elif 1 // FURTHEST
+                uint16_t idx = *indices++;
+                sortScratch[j].fIdx[0] = idx;
+                hsPoint3 pos = *(hsPoint3*)(vdata + idx * stride);
+                float dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                float maxDist = dist;
+
+                idx = *indices++;
+                sortScratch[j].fIdx[1] = idx;
+                pos = *(hsPoint3*)(vdata + idx * stride);
+                dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                if( dist > maxDist )
+                    maxDist = dist;
+
+                idx = *indices++;
+                sortScratch[j].fIdx[2] = idx;
+                pos = *(hsPoint3*)(vdata + idx * stride);
+                dist = hsVector3(&pos, &viewPos).MagnitudeSquared();
+                if( dist > maxDist )
+                    maxDist = dist;
+
+                sortScratch[j].fDist = maxDist;
+#endif // SORTTYPES
+            }
+
+            std::sort(sortScratch.begin(), sortScratch.end(), plCompSortFace());
+
+            indices = group->GetIndexBufferData(span->fIBufferIdx) + span->fIStartIdx;
+            for (const plSortFace& iter : sortScratch)
+            {
+                *indices++ = iter.fIdx[0];
+                *indices++ = iter.fIdx[1];
+                *indices++ = iter.fIdx[2];
+            }
+
+            group->DirtyIndexBuffer(span->fIBufferIdx);
+        }
+    }
+    return true;
+}
+
+template <class DeviceType>
+bool pl3DPipeline<DeviceType>::ISoftwareVertexBlend(plDrawableSpans* drawable, const std::vector<int16_t>& visList)
+{
+    if (IsDebugFlagSet(plPipeDbg::kFlagNoSkinning))
+        return true;
+
+    if (drawable->GetSkinTime() == fRenderCnt)
+        return true;
+
+    const hsBitVector& blendBits = drawable->GetBlendingSpanVector();
+
+    if (drawable->GetBlendingSpanVector().Empty()) {
+        // This sucker doesn't have any skinning spans anyway. Just return
+        drawable->SetSkinTime(fRenderCnt);
+        return true;
+    }
+
+    plProfile_BeginTiming(Skin);
+
+    // First, figure out which buffers we need to blend.
+    constexpr size_t kMaxBufferGroups = 20;
+    constexpr size_t kMaxVertexBuffers = 20;
+    static char blendBuffers[kMaxBufferGroups][kMaxVertexBuffers];
+    memset(blendBuffers, 0, kMaxBufferGroups * kMaxVertexBuffers * sizeof(**blendBuffers));
+
+    hsAssert(kMaxBufferGroups >= drawable->GetNumBufferGroups(), "Bigger than we counted on num groups skin.");
+
+    const std::vector<plSpan*>& spans = drawable->GetSpanArray();
+    for (int16_t idx : visList) {
+        if (blendBits.IsBitSet(idx)) {
+            const plVertexSpan& vSpan = *reinterpret_cast<plVertexSpan*>(spans[idx]);
+            hsAssert(kMaxVertexBuffers > vSpan.fVBufferIdx, "Bigger than we counted on num buffers skin.");
+
+            blendBuffers[vSpan.fGroupIdx][vSpan.fVBufferIdx] = 1;
+            drawable->SetBlendingSpanVectorBit(idx, false);
+        }
+    }
+
+    // Now go through each of the group/buffer (= a real vertex buffer) pairs we found,
+    // and blend into it. We'll lock the buffer once, and then for each span that
+    // uses it, set the matrix palette and and then do the blend for that span.
+    // When we've done all the spans for a group/buffer, we unlock it and move on.
+    for (size_t i = 0; i < kMaxBufferGroups; i++) {
+        for (size_t j = 0; j < kMaxVertexBuffers; j++) {
+            if (blendBuffers[i][j]) {
+                // Found one
+                typename DeviceType::VertexBufferRef* vRef = reinterpret_cast<typename DeviceType::VertexBufferRef*>(drawable->GetVertexRef(i, j));
+                hsAssert(vRef->fData, "Going into skinning with no place to put results!");
+
+                uint8_t* destPtr = vRef->fData;
+
+                for (int16_t idx : visList) {
+                    const plIcicle& span = *reinterpret_cast<plIcicle*>(spans[idx]);
+                    if ((span.fGroupIdx == i) && (span.fVBufferIdx == j)) {
+                        plProfile_Inc(NumSkin);
+
+                        hsMatrix44* matrixPalette = drawable->GetMatrixPalette(span.fBaseMatrix);
+                        matrixPalette[0] = span.fLocalToWorld;
+
+                        uint8_t* ptr = vRef->fOwner->GetVertBufferData(vRef->fIndex);
+                        ptr += span.fVStartIdx * vRef->fOwner->GetVertexSize();
+                        IBlendVertsIntoBuffer(&span,
+                                matrixPalette, span.fNumMatrices,
+                                ptr,
+                                vRef->fOwner->GetVertexFormat(),
+                                vRef->fOwner->GetVertexSize(),
+                                destPtr + span.fVStartIdx * vRef->fVertexSize,
+                                vRef->fVertexSize,
+                                span.fVLength,
+                                span.fLocalUVWChans);
+                        vRef->SetDirty(true);
+                    }
+                }
+            }
+        }
+    }
+
+    plProfile_EndTiming(Skin);
+
+    if (drawable->GetBlendingSpanVector().Empty()) {
+        // Only do this if we've blended ALL of the spans. Thus, this becomes a trivial
+        // rejection for all the skinning flags being cleared
+        drawable->SetSkinTime(fRenderCnt);
+    }
+
+    return true;
 }
 
 #endif //_pl3DPipeline_inc_
